% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{corporate}
\alias{corporate}
\title{Turn a dataframe of papers into a corpus}
\usage{
corporate(df, col = "AB")
}
\arguments{
\item{df}{A dataframe of papers}
}
\value{
A corpus of words, punct removed and stemmed
}
\description{
Turn a dataframe of papers into a corpus
}

