% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gRate}
\alias{gRate}
\title{Apply OECD categories to web of science subject area}
\usage{
gRate(df, field, y1, y2, total = F)
}
\arguments{
\item{df}{A dataframe of papers}

\item{field}{The field to group growth rates by}

\item{y1}{The year to start from}

\item{y2}{The year to finish}

\item{total}{Is the df already summarised (default=F)}
}
\value{
A table of growth rates by field
}
\description{
Apply OECD categories to web of science subject area
}
