% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{makeDTM}
\alias{makeDTM}
\title{Turn a corpus into a document term matrix}
\usage{
makeDTM(corpus, sparsity, rnames, cols, rows)
}
\arguments{
\item{df}{A dataframe of papers}
}
\value{
A corpus of words, punct removed and stemmed
}
\description{
Turn a corpus into a document term matrix
}

