% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoS.R
\name{mergeOECD}
\alias{mergeOECD}
\title{Apply OECD categories to web of science subject area}
\usage{
mergeOECD(df)
}
\arguments{
\item{df}{A dataframe of papers}
}
\value{
A dataframe of WoS records with OECD fields
}
\description{
Apply OECD categories to web of science subject area
}

