% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{paperNumbers}
\alias{paperNumbers}
\title{Plot numbers of papers by variable and year}
\usage{
paperNumbers(df, field, n = "all", summed = F, graph = F, bSize = 12)
}
\arguments{
\item{df}{a dataframe of papers}

\item{field}{a string describing the variable to pass to the bar fill}

\item{n}{the number of categories to display (defaults to "all")}

\item{summed}{is the data already summed? (default F)}

\item{graph}{return a graph object (default FALSE) or data}

\item{bSize}{set the base size of the text}
}
\value{
Either a data frame of year and category sums, or a graph object
}
\description{
Plot numbers of papers by variable and year
}

