% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{paperShares}
\alias{paperShares}
\title{Plot shares of papers by variable and year}
\usage{
paperShares(df, field, n = "all", pType = "bar", bSize = 12)
}
\arguments{
\item{df}{a dataframe of papers}

\item{field}{a string describing the variable to pass to the bar fill}

\item{n}{the number of categories to display (defaults to "all")}

\item{pType}{plot type, either "bar" or "line"}

\item{bSize}{set the base size of the text}
}
\value{
A dataframe of the summed results
}
\description{
Plot shares of papers by variable and year
}

