% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_k.R
\name{plot.optimal_k1}
\alias{plot.optimal_k1}
\title{Plots a plot.optimal_k1 Object}
\usage{
\method{plot}{plot.optimal_k1}(x, ...)
}
\arguments{
\item{x}{A \code{optimal_k1} object.}

\item{\ldots}{Ignored.}
}
\description{
Plots a plot.optimal_k1 object
}

