% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_k.R
\name{plot.optimal_k2}
\alias{plot.optimal_k2}
\title{Plots a plot.optimal_k2 Object}
\usage{
\method{plot}{plot.optimal_k2}(x, ...)
}
\arguments{
\item{x}{A \code{optimal_k2} object.}

\item{\ldots}{Ignored.}
}
\description{
Plots a plot.optimal_k2 object
}

