% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_biblioAnalysis}
\alias{plot_biblioAnalysis}
\title{Plot a summary of the results from the bilbiometric analysis}
\usage{
plot_biblioAnalysis(df)
}
\arguments{
\item{df}{a dataframe of papers}
}
\description{
Plot a summary of the results from the bilbiometric analysis
}
\author{
Jerome Hilaire
}

