% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_k.R
\name{print.optimal_k}
\alias{print.optimal_k}
\title{Prints a optimal_k Object}
\usage{
\method{print}{optimal_k}(x, ...)
}
\arguments{
\item{x}{A \code{optimal_k} object.}

\item{\ldots}{Ignored.}
}
\description{
Prints a optimal_k object
}
