% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoS.R
\name{readWoS}
\alias{readWoS}
\title{Read data from Web of Science (WoS)}
\usage{
readWoS(p)
}
\arguments{
\item{p}{A path to a WoS file or directory of WoS files}
}
\value{
A dataframe of WoS records
}
\description{
Read data from Web of Science (WoS)
}

