% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{refresh_corp}
\alias{refresh_corp}
\title{Make sure the corpus reflects the dtm after working on it}
\usage{
refresh_corp(dtm)
}
\arguments{
\item{dtm}{a document term matrix}
}
\value{
A corpus of words that reflect the terms in the dtm
}
\description{
Make sure the corpus reflects the dtm after working on it
}

