% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sumKeywords}
\alias{sumKeywords}
\title{Calculate and plot keyword sums}
\usage{
sumKeywords(df, n = 100, bSize = 12)
}
\arguments{
\item{df}{a dataframe of papers}

\item{n}{the number of kewords to display (defaults to 100)}

\item{bSize}{set the base size of the text}
}
\value{
A dataframe of the summed results
}
\description{
Calculate and plot keyword sums
}

