% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{topCors}
\alias{topCors}
\title{Produce a topic correlation graph file}
\usage{
topCors(model, topic_names = FALSE)
}
\arguments{
\item{model}{a topic model}

\item{topic_names}{a character vector of topic names}
}
\description{
Produce a topic correlation graph file
}
