% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_ldavis.R
\name{topicmodels_json_ldavis}
\alias{topicmodels_json_ldavis}
\title{Output topic model results in json format}
\usage{
topicmodels_json_ldavis(fitted, corpus, doc_term)
}
\arguments{
\item{fitted}{a topic model}

\item{corpus}{a corpus}

\item{doc_term}{a document term matrix}
}
\description{
Output topic model results in json format
}
