% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{makeDTM}
\alias{makeDTM}
\title{Turn a corpus into a document term matrix}
\usage{
makeDTM(corpus, sparsity, rnames, cols, rows)
}
\arguments{
\item{corpus}{- A document corpus}

\item{sparsity}{- An integer between 0 and 1. Lower values remove more highly common and highly uncommon terms}

\item{rnames}{- a vector of document identifiers (usually papers$UT)}
}
\value{
A corpus of words, punct removed and stemmed
}
\description{
Turn a corpus into a document term matrix
}

