% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{topCors}
\alias{topCors}
\title{Produce a topic correlation graph file}
\usage{
topCors(model)
}
\arguments{
\item{model}{a topic model}

\item{corpus}{a corpus}

\item{dtm}{a document term matrix}
}
\description{
Produce a topic correlation graph file
}

