% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{visualise}
\alias{visualise}
\title{visualise the topic model}
\usage{
visualise(model, corpus, dtm, dir = F)
}
\arguments{
\item{model}{a topic model}

\item{corpus}{a corpus}

\item{dtm}{a document term matrix}

\item{dir}{(optional - defaults to model name) a directory to save the results}
}
\description{
visualise the topic model
}

